Bring in the standalone midrd daemon and group 2's single-instance
flooding so group 1 can move onto the midrd topology API.

Conflicts were limited to group-1 files in bgpd, where both branches
carried a version of IPv6 NDS support.  Our side keeps the newer IPv6
integration and Tier1 admission; colleague-only group-1 changes were
carried over by hand:
- 88d2193ba1 skip conflicted bootstrap candidates (applied as is)
- 200245a270 graceful MIDR withdrawal (ported to ipaddr transports)
- bf20cfd4fa unused variables (already absent on our side)

verify_group1_group2_evidence.py takes group 2's version, which no
longer checks group sizes or propagation paths.

+++ lib/id_alloc.c
@@ -127,7 +127,7 @@ void idalloc_free(struct id_alloc *alloc, uint32_t id)
 	word = ID_WORD(id);
 	offset = ID_OFFSET(id);
 
-	if ((page->allocated_mask[word] & (1 << offset)) == 0) {
+	if ((page->allocated_mask[word] & ((uint32_t)1 << offset)) == 0) {
 		flog_err(EC_LIB_ID_CONSISTENCY,
 			"ID Allocator %s cannot free #%u. ID was not allocated at the time of free.",
 			alloc->name, id);
